+++ CHANGELOG.md
@@ -10,10 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `PayIdClient` provides the functionality of `PayIDClient` under an idiomatic cased name.
 - `PayIdUtils` provides the functionality of `PayIDUtils` under an idiomatic cased name.
+- `XrpPayIdClient` provides the functionality of `XRPPayIDClient` under an idiomatic cased name.
 
 ### Deprecated
 - `PayIDClient` is deprecated, use the idiomatically named `PayIdClient` instead.
 - `PayIDUtils` is deprecated, use the idiomatically named `PayIdUtils` instead.
+- `XRPPayIDClient` is deprecated, use the idiomatically named `XrpPayIdClient` instead.
 
 ## 5.2.0 - 2020-06-01