+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
+
+## 5.2.1 - 2020-06-16
+
 ### Added
 - `XrpPayment` and `XrpTransaction` now contain X-address representations of their address and tag fields.
   (See https://xrpaddress.info/)